Refactor mapper tests into package
