+++ nes_py/tests/mappers/__init__.py
@@ -1,0 +1,1 @@
+"""Mapper registry, lifecycle, and per-board characterization tests."""
