+++ nes_py/tests/test_multiple_makes.py
@@ -2,7 +2,7 @@
 from multiprocessing import Process
 from threading import Thread
 from unittest import TestCase
-from .rom_file_abs_path import rom_file_abs_path
+from nes_py.tests.rom_file_abs_path import rom_file_abs_path
 from nes_py.nes_env import NESEnv
