+++ nes_py/tests/test_rom.py
@@ -5,7 +5,7 @@ http://tuxnes.sourceforge.net/nesmapper.txt
 
 """
 from unittest import TestCase
-from .rom_file_abs_path import rom_file_abs_path
+from nes_py.tests.rom_file_abs_path import rom_file_abs_path
 from nes_py._rom import ROM
